requirements.txt
django-payment[mercadopago]